+++ src/multicall.ts
@@ -49,6 +49,7 @@ export const multicallAddresses: { [ChainId: number]: string } = {
   [ChainId.OPB]: '0xcA11bde05977b3631167028862bE2a173976CA11',
   [ChainId.WCC]: '0xcA11bde05977b3631167028862bE2a173976CA11',
   [ChainId.LSK]: '0xcA11bde05977b3631167028862bE2a173976CA11',
+  [ChainId.ABS]: '0xAa4De41dba0Ca5dCBb288b7cC6b708F3aaC759E7',
   [ChainId.SON]: '0xcA11bde05977b3631167028862bE2a173976CA11',
 
   // TODO